refactor(local-mail): share one engine and make the HTTP surface mountable
Two hosts need to open every connected mailbox and keep it synced (ADR-0191),
so the part that does it moves out of the standalone host and into
`src/engine.ts`: account selection, sync sessions, per-account gates, sync
locks, and the poll loop. `runApp` keeps what is actually its own, the
loopback bind, the Host-check kill switch, the bearer, and static SPA serving,
and drops about a third of its body. Neither host can now drift from the other
on what an open account means, because there is only one answer.

The HTTP surface loses its `/api` prefix and its bearer middleware. Both were
the standalone host's, not the surface's: they are why the routes could not be
mounted anywhere else, and why they would have collided with Epicenter's own
`/api/apps` and `/api/home/session`. `runApp` now applies both itself, which
is where the Host check already lived.

`close()` becomes async and drains each account's gate before closing its
session. Aborting and closing in the same tick left an in-flight sync pass
running against finalized statements, which the compiled-host test surfaced as
"Statement has finalized". The poll sleep is now abort-aware too, so shutdown
never waits out a full interval.

MailUnavailable joins the error family for a host that mounted the routes with
no engine behind them. It is the mail surface's vocabulary because the SPA's
one error path reads it, even though only a host emits it.

The bearer-gate test goes away with the gate it covered. Rebuilding it against
a copy of the host's composition would assert the copy, not the host; the gate
is eight lines with a scheduled deletion, and this records that it is now
uncovered.
